LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/items/row-minimum-size-grid-items-002-expected.txt


FAIL .grid-lanes 1 assert_equals:
<div class="grid-lanes" style="grid-template-rows: auto;">
  <div class="verticalLR" data-expected-height="100">XXXXXXXXXX</div>
  <div data-expected-height="100"></div>
</div>
height expected 100 but got 0
PASS .grid-lanes 2
FAIL .grid-lanes 3 assert_equals:
<div class="grid-lanes" style="grid-template-rows: 25px;">
  <div class="verticalLR" data-expected-height="25">XXXXXXXXXX</div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
PASS .grid-lanes 4
FAIL .grid-lanes 5 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalLR" data-expected-height="25">XXXXXXXXXX</div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 6 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 7 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalLR" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 8 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="0" style="margin: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 9 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalLR" data-expected-height="15" style="margin: 5px 0px;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 10 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="10" style="padding: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 11 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalLR" data-expected-height="25" style="padding: 5px 0px;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 12 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="10" style="border: solid 5px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 13 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalLR" data-expected-height="25" style="border: solid 5px blue;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 14 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalLR" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 15 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalLR" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 16 assert_equals:
<div class="grid-lanes" style="grid-template-rows: auto;">
  <div class="verticalRL" data-expected-height="100">XXXXXXXXXX</div>
  <div data-expected-height="100"></div>
</div>
height expected 100 but got 0
PASS .grid-lanes 17
FAIL .grid-lanes 18 assert_equals:
<div class="grid-lanes" style="grid-template-rows: 25px;">
  <div class="verticalRL" data-expected-height="25">XXXXXXXXXX</div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
PASS .grid-lanes 19
FAIL .grid-lanes 20 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalRL" data-expected-height="25">XXXXXXXXXX</div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 21 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 22 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalRL" data-expected-height="10" style="height: 10px;">XXXXXXXXXX</div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 23 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="0" style="margin: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 24 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalRL" data-expected-height="15" style="margin: 5px 0px;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 25 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="10" style="padding: 5px 0px;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 26 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalRL" data-expected-height="25" style="padding: 5px 0px;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 27 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="10" style="border: solid 5px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 28 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalRL" data-expected-height="25" style="border: solid 5px blue;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 29 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 0px);">
  <div class="verticalRL" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="10"></div>
</div>
height expected 10 but got 0
FAIL .grid-lanes 30 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(auto, 25px);">
  <div class="verticalRL" data-expected-height="8" style="height: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-height="25"></div>
</div>
height expected 25 but got 0
FAIL .grid-lanes 31 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: auto;">
  <div class="horizontalTB" data-expected-width="100">XXXXXXXXXX</div>
  <div data-expected-width="100"></div>
</div>
width expected 100 but got 0
PASS .grid-lanes 32
FAIL .grid-lanes 33 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: 25px;">
  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
PASS .grid-lanes 34
FAIL .grid-lanes 35 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 36 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 37 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 38 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="0" style="margin: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 39 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="15" style="margin: 0px 5px;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 40 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="padding: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 41 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="25" style="padding: 0px 5px;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 42 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="border: solid 5px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 43 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="25" style="border: solid 5px blue;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 44 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 45 assert_equals:
<div class="grid-lanes verticalLR" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 46 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: auto;">
  <div class="horizontalTB" data-expected-width="100">XXXXXXXXXX</div>
  <div data-expected-width="100"></div>
</div>
width expected 100 but got 0
PASS .grid-lanes 47
FAIL .grid-lanes 48 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: 25px;">
  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
PASS .grid-lanes 49
FAIL .grid-lanes 50 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="25">XXXXXXXXXX</div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 51 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 52 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="10" style="width: 10px;">XXXXXXXXXX</div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 53 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="0" style="margin: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 54 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="15" style="margin: 0px 5px;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 55 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="padding: 0px 5px;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 56 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="25" style="padding: 0px 5px;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 57 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="10" style="border: solid 5px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 58 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="25" style="border: solid 5px blue;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
FAIL .grid-lanes 59 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 0px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="10"></div>
</div>
width expected 10 but got 0
FAIL .grid-lanes 60 assert_equals:
<div class="grid-lanes verticalRL" style="grid-template-rows: minmax(auto, 25px);">
  <div class="horizontalTB" data-expected-width="8" style="width: 4px; margin: 1px; padding: 1px; border: solid 1px blue;"></div>
  <div data-expected-width="25"></div>
</div>
width expected 25 but got 0
item writing-mode: vertical-lr;

grid-template-rows: auto;
XXXXXXXXXX
grid-template-rows: 0px;
XXXXXXXXXX
grid-template-rows: 25px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item margin height: 10px;
grid-template-rows: minmax(auto, 25px); item margin height: 10px;
grid-template-rows: minmax(auto, 0px); item padding height: 10px;
grid-template-rows: minmax(auto, 25px); item padding height: 10px;
grid-template-rows: minmax(auto, 0px); item border height: 10px;
grid-template-rows: minmax(auto, 25px); item border height: 10px;
grid-template-rows: minmax(auto, 0px); item height + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 25px); item height + margin + border + padding: 10px;
item writing-mode: vertical-rl;

grid-template-rows: auto;
XXXXXXXXXX
grid-template-rows: 0px;
XXXXXXXXXX
grid-template-rows: 25px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px); item height: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item margin height: 10px;
grid-template-rows: minmax(auto, 25px); item margin height: 10px;
grid-template-rows: minmax(auto, 0px); item padding height: 10px;
grid-template-rows: minmax(auto, 25px); item padding height: 10px;
grid-template-rows: minmax(auto, 0px); item border height: 10px;
grid-template-rows: minmax(auto, 25px); item border height: 10px;
grid-template-rows: minmax(auto, 0px); item height + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 25px); item height + margin + border + padding: 10px;
masonry container writing-mode: vertical-lr; & item writing-mode: horizontal-tb;

grid-template-rows: auto;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: 25px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item margin width: 10px;
grid-template-rows: minmax(auto, 25px); item margin width: 10px;
grid-template-rows: minmax(auto, 0px); item padding width: 10px;
grid-template-rows: minmax(auto, 25px); item padding width: 10px;
grid-template-rows: minmax(auto, 0px); item border width: 10px;
grid-template-rows: minmax(auto, 25px); item border width: 10px;
grid-template-rows: minmax(auto, 0px); item width + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 25px); item width + margin + border + padding: 10px;
masonry container writing-mode: vertical-rl; & item writing-mode: horizontal-tb;

grid-template-rows: auto;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: 25px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px);
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 25px); item width: 10px;
XXXXXXXXXX
grid-template-rows: minmax(auto, 0px); item margin width: 10px;
grid-template-rows: minmax(auto, 25px); item margin width: 10px;
grid-template-rows: minmax(auto, 0px); item padding width: 10px;
grid-template-rows: minmax(auto, 25px); item padding width: 10px;
grid-template-rows: minmax(auto, 0px); item border width: 10px;
grid-template-rows: minmax(auto, 25px); item border width: 10px;
grid-template-rows: minmax(auto, 0px); item width + margin + border + padding: 10px;
grid-template-rows: minmax(auto, 25px); item width + margin + border + padding: 10px;
